Fixed a bug in the way in which PreEvaluationCheck was computing tile sizes. Projecting an exclusive operation point onto data-space points is erroneous. Only inclusive points should be projected, and an exclusive point in the data-space should be computed thereafter.
Fixed a bug in mapper diagnostics. If all failed mappings were rejected by PreEvaluationCheck, then the diagnostics won't have any post-evaluation stats to print out. To address this, the diagnostics code now runs an Evaluate() call on the sample mapping to be printed if it wasn't evaluated on the engine already.