[CompoundConfig] A new wrapper class to support libconfig and yaml
This new class is an attempt to replace all (or at least, most of)
libconfig dependency in Timeloop. The goal to ultimately let the yaml
file used in other tools (Sym-Sim, Accelergy) shared and used as inputs
in Timeloop for further co-optimization. For example, we might want to
do memory hierarchy design space and loop tiling co-optimization.

This commit is a first attemp by providing the neccesary APIs for
Timeloop. Only

applications/mapper/main.cpp
applications/mapper/mapper.hpp

has adopted the new APIs.

The next step is to replace libconfig calls in other files and
perform a regression test to make sure the original functionality is not
affected. Then following that, we will implement the yaml support within
the CompoundConfig class.