[CompoundConfig] Fix bugs for yaml support
Bug 1: lookupValue() should return false when the value to look up is
not a scalar, but a list or map.

Bug 2: isList() should return true is the list is empty. If it is not
empty, then we will use the type of the first element to decide if it is
a list or an array

Also, given that the yaml-cpp libarary does not show where the error
shows up in the input file in the exception messeage, it's better not to
capture it and fails on the spot.