[CompoundConfig] Handle yaml config error better
yaml-cpp does not have a good implementation of error handling. For
example, KeyNotFound exception exists but is never used by any API.
Also, any failed lookup will not throw a exception but return an invalid
node, and when we try to do something with the invalid node, it throws a
InvalidNode exception at that spot, which is not helpful for debugging
and cause user frustrations.

As an attempt to address, the lookup function in CompoundConfigNode now
check if the Ynode is valid and will throw "KeyNotFound" exception if
the key is not valid.