[Buffer,CompoundConfigNode] Parse integer when expecting float
Originally, libconfig needs to parse both int and double/float
separately, and therefore we have multiple code block to handle this.
Now that we have the CompoundConfigNode wrapper, we can perfom this
logic once in the lookupValue function.

YAML config also has a fallback type support for the as() function.
Originally, we use "as<double>()"; now for all value expecting
double/float, we use "as<double, int>(0)" to parse integer as floating
point value.

Note that only int -> float/double is okay. Passing float/double to int
should still throw errors because we should not guess how the user
would like the value to be rounded.