[scons] Add --accelergy flag to build Timeloop with Accelergy
With this flag, during compilation, scons will search $PATH to find
where Accelergy is installed and pass it to Timeloop. And if scons
cannot find Accelergy, the build will fail.

The ACCELERGYPATH variable now becomes an option to provide only if
Accelergy is not installed as instructed (i.e., `which accelergy` cannot
find it).