Significant refactoring of network code. Network modules are now direct children of the topology module. At the moment, there is a 1:1 correspondence between a network instance and a storage level instance. The topology module (and the XML stat parser) knows about this correspondence and explicitly uses it. We need to have a more general solution for this.
Removed the notion of Fanout from the network architecture specs. We were using Fanout to capture the maximum spatial fanout allowed by the hardware between two tiling levels. We have been thinking of an abstract/logical architectural topology that the mapper targets, which is later "bound" to the physical hardware topology. Fanout appears to belong to that abstract level. The existing ArchProperties class appears to be a good starting point for this abstract architecture level, and so we moved the Fanout property to this class. The Mapspace now instantiates an ArchProperties object and queries it in various places. It no longer interacts directly with the physical architecture specs. Note that fanout still exists as a *stat* within the physical network class, and that seems to be okay.

Added a debug flag to the SConstruct.

After the refactoring, Arithmetic units will need to explicitly specificy MeshX/Y *if* they are also specified for other storage levels.

Access counts match pre-refactor values. However, energy calculations are off. This needs to be fixed.