[Mapper] Make the mapper binary produce the best mapping in yaml
When use yaml inputs, the mapper binary (timeloop-mapper) now produces
an output yaml, OUTPUT_PREFIX.map.yaml, which can be used as an input
yaml for timeloop-model.

This feature was only avaiable for .cfg inputs, and now the mapper will
produce the best mapping with either .cfg or .yaml inputs. Note that it
still assumes only *one* format is used at a time.