Remove problem::GetShape() references from files in model/.
This required us to store a workload_ pointer in each module (e.g.,
Topology, Buffer, all network modules) which would be set at the
beginning of an Evaluate() call, and subsequently used by various
other methods in each class. This partial persistence of a legal
workload_ pointer is ugly and we should try to find an alternative
solution.

Another issue was that we were using a convention for stat
accessors (such as buffer->Energy(dataspace_id)) with a default
parameter of dataspace_id == NumDataSpaces that returned
accumulated stats across all data spaces (or dataspace-independent
stats if applicable). We had to remove the default parameter for
these calls, and instead explicitly call the methods with
NumDataSpaces where applicable. Fortunately all of these calls
were in the Topology class and therefore not visible to the model
or mapper applications. An alternative would have been to create
two methods for each accessor, one with the dataspace_id parameter
and another without.

As of this commit, the only remaining global problem::GetShape()
references are in the sparse analysis code.