Translation fix (#297)
* The method OperationSpace::SaveAndSubtractIfSameStride() is called by the
classic NestAnalysis code. Its job is to determine how the translation
vector (or stride vector) between dataspace point-sets changes between
consecutive loop iterations. Note that if the translation vector is the
"first-order derivative" of point sets, then this computation is effectively
the second-order derivative.

If this translation vector _changes_, then this code avoids computing the
delta between consecutive point-sets and instead discards the old point set
and installs the new one. This is done to emulate the behavior of simplistic
hardware pattern generators that are unable to describe hierarchical patterns.

However, this code needs a number of special cases, particularly stationarity
detection (translation == 0). This is because we do want to allow patterns
such as translation=<N,0,0,0,N,0,0,0,...>. However, this specific case
was being detected as a translation vector change, resulting in the old
point-set being thrown away and increasing the number of accesses to the next
storage level.

The entire logic has been re-written to carefully handle all cases.

* Add a comment with the full case table for
OperationSpace::SaveAndSubtractIfSameStride().